requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==3.0.11
django-allauth==0.50.0
django-cors-headers==3.6.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.2.0
platformdirs==2.5.1
psutil==5.9.1
psycopg2==2.8.6
pycparser==2.21
PyJWT==1.7.1
pyotp==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-env==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.13.3
Werkzeug==2.1.2
zipp==3.8.0